requirements.txt
# Core MCP and server dependencies
fastmcp>=0.1.0

# Vector database
qdrant-client>=1.7.0

# Embedding models
sentence-transformers>=2.2.0
torch>=2.0.0

# Numerical computing
numpy>=1.24.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# WikiBooks crawler dependencies
mwparserfromhell>=0.6.0
requests>=2.31.0

# Optional: For better performance
accelerate>=0.20.0
